+++ include/typing/serializer.h
@@ -19,7 +19,7 @@ struct Serializer {
         [&](auto&& arg) {
           using T = std::decay_t<decltype(arg)>;
 
-          if constexpr (std::is_same_v<T, int>) {
+          if constexpr (std::is_same_v<T, long long>) {
             type = ValueType::Integer;
           } else if constexpr (std::is_same_v<T, double>) {
             type = ValueType::Double;
@@ -82,7 +82,7 @@ struct Serializer {
 
     switch (get_value_type(v)) {
       case ValueType::Integer:
-        sv << std::get<int>(v);
+        sv << std::get<long long>(v);
         break;
       case ValueType::Double:
         sv << std::get<double>(v);
