+++ include/typing/valuetype.h
@@ -25,7 +25,7 @@ enum class ValueType {
   Lambda
 };
 
-using Value = std::variant<int, double, bool, std::string,
+using Value = std::variant<long long, double, bool, std::string,
                            std::shared_ptr<List>, std::shared_ptr<Hash>,
                            std::shared_ptr<Object>, std::shared_ptr<LambdaRef>>;
